Setup Wizard - Detect current WiFi IP, prompt for broker
During Setup, if connected to Sengled already, it tends to detect/use 127.0.0.1 as the server address because it isn't aware of the local WiFi network IP. To fix this, detect the IP before prompting to switch to the Sengled WiFi setup network. Messaging can still be improved.

Also prompt for MQTT broker IP, which may not be the local PC. Defaults to local IP if skipped.