Improve WiFi setup for internal server instance
The setup -> upgrade workflow (and indeed, smooth operation of Sengled-Tools itself) requires you to keep an instance of the embedded servers in the background, so the bulb can stay connected reliably (if not connected at the time of an MQTT command, it'll miss the command).

So this is a rework to allow "--run-servers" to run in one window, and run setup and upgrade commands in another. Ran through 6 bulbs tonight to reflash Tasmota after a long session of crafting interaction techniques to detect, use, and react to state of the embedded server in the other instance.